🪡 :: move out configure.py, optimizer.py
